+++ src/blakron/components/Group.ts
@@ -1,4 +1,4 @@
-import { Sprite, Rectangle, Point, DisplayObject, Event, type DisplayObjectEvents } from '@blakron/core';
+import { Sprite, Rectangle, Point, DisplayObject, Event, type DisplayObjectEvents, type Matrix } from '@blakron/core';
 import { UIState, isUIComponent } from '../core/UIState.js';
 import type { IUIOwner } from '../core/UIState.js';
 import type { IUIComponent } from '../core/IUIComponent.js';
@@ -228,39 +228,44 @@ export class Group extends Sprite implements IUIComponent, IViewport, ILayoutTar
 		this.ui.percentHeight = v;
 	}
 
-	public override get scaleX(): number { return super.scaleX; }
-	public override set scaleX(v: number) {
-		if (super.scaleX === v) return;
-		super.scaleX = v;
-		this.ui._invalidateParentLayout();
+	public override $updateUseTransform(): void {
+		super.$updateUseTransform();
+		this.ui.$invalidateParentLayout();
 	}
 
-	public override get scaleY(): number { return super.scaleY; }
-	public override set scaleY(v: number) {
-		if (super.scaleY === v) return;
-		super.scaleY = v;
-		this.ui._invalidateParentLayout();
+	public override $setMatrix(matrix: Matrix, needUpdateProperties = true): void {
+		super.$setMatrix(matrix, needUpdateProperties);
+		this.ui.$invalidateParentLayout();
 	}
 
-	public override get rotation(): number { return super.rotation; }
-	public override set rotation(v: number) {
-		if (super.rotation === v) return;
-		super.rotation = v;
-		this.ui._invalidateParentLayout();
+	public override $setAnchorOffsetX(value: number): void {
+		if (this.$anchorOffsetX === value) return;
+		super.$setAnchorOffsetX(value);
+		this.ui.$invalidateParentLayout();
 	}
 
-	public override get skewX(): number { return super.skewX; }
-	public override set skewX(v: number) {
-		if (super.skewX === v) return;
-		super.skewX = v;
-		this.ui._invalidateParentLayout();
+	public override $setAnchorOffsetY(value: number): void {
+		if (this.$anchorOffsetY === value) return;
+		super.$setAnchorOffsetY(value);
+		this.ui.$invalidateParentLayout();
 	}
 
-	public override get skewY(): number { return super.skewY; }
-	public override set skewY(v: number) {
-		if (super.skewY === v) return;
-		super.skewY = v;
-		this.ui._invalidateParentLayout();
+	public override $setX(value: number): boolean {
+		const changed = super.$setX(value);
+		if (changed) {
+			this.ui.$invalidateParentLayout();
+			this.invalidateProperties();
+		}
+		return changed;
+	}
+
+	public override $setY(value: number): boolean {
+		const changed = super.$setY(value);
+		if (changed) {
+			this.ui.$invalidateParentLayout();
+			this.invalidateProperties();
+		}
+		return changed;
 	}
 
 	public override get width(): number {